refactor: move ExpenseCategory and rename it to ExpenseCategoryModel
